made test for apply pipeline and updated pipeline unittests
