+++ arcana/data/stores/common/tests/test_file_system.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import operator as op
 from functools import reduce
 from arcana.core.data.set import Dataset
-from arcana.tests.fixtures.common import create_test_file
+from arcana.test.datasets import create_test_file
