--- arcana/tests/utils.py
@@ -1,4 +1,0 @@
-import traceback
-
-def show_cli_trace(result):
-    return ''.join(traceback.format_exception(*result.exc_info))
